refactory bulk load
