+++ bin/README.md
@@ -1,0 +1,10 @@
+#FILES
+- `all_variables.json`: This is the output of https://github.com/chicagopcdc/pcdcmodels. #TODO We should pull automatically the latest list of variables instead of having this file here, but that repo is not fully ready yet.
+- `bulk_load_terms_file.py`: This is loading scrolling all the codes in the `all_variables.json` file and pulling from the NCIt website the information connected to the codes and populating the `_terms.yaml` file in the DD. This is an expensive operation the first time but it should get faster overtime when the terminology will be more complete.
+- `create_json.py`: This is used to dump the YAML in a schema.json format for GEN3
+- `dump_schem.py`: deprecated version of `create_json.py` #TODO, we may remove this
+- `load_aws.sh`: This is using the CLI interface to load the schema to the desired S3 bucket. It requires the AWS CLI to be installed and the profile to be logged in on the machine it is running. #TODO needs some updates and configs to avoid changing the codes for different users. Haven't used in a while may be broken with most up to date CLI version.
+- `utils.py`: This contains functions reused in the other files.
+
+#DEPENDENCIES
+The dictionaryutils repo dependencies need to be installed via poetry install.
